requirements.txt
pandas==1.3.4
Pillow==9.4.0
streamlit==1.17.0
